implement a GitLab title grabber for Tor
This implements a crude GitlabTitleProvider based on the
TicketHtmlTitleProvider. It's a little hackish because it requires
extra backflips to take into account the "path" section of the
shortcut.

This should parse patterns like PATH#NUMBER and turn those into
https://gitlab.torproject.org/PATH/-/issues/NUMBER URLs. Then it will
extract the title from that page as normal *but* it will add the URL
as a postfix manually, instead of relying on 'postfix' because that
assumes we only have one variable component in the URL.

That required, in turn, hacking at the _gittitle function but that was
done in a backwards-compatible way, so that should work.